refactor elevator class into seperate header/cpp files
