+++ elevator.h
@@ -1,0 +1,20 @@
+#ifndef ELEVATOR_H
+#define ELEVATOR_H
+#include <vector>
+
+constexpr int DEFAULT_FLOOR_TRAVEL_TIME { 10 };
+
+class Elevator 
+{
+private:
+    int floor {};
+    int floorTravelTime {};
+    int totalTravelTime { 0 };
+    std::vector<int> visitedFloors {};
+public:
+    Elevator(int initialFloor, int initialFloorTravelTime = DEFAULT_FLOOR_TRAVEL_TIME);
+    void travelToFloor(int destinationFloor);
+    void printTravelTime();
+    void printTraversedFloors();
+};
+#endif
